Separate Bob restart tests
To stabilize CI
